Make sure script-based button behaviours match that of Mixxx...
...we added a series of hacks during the development of the Xponent that
do not match the reality of Mixxx.  One of them is that note offs are
translated to 0.  The second one is that binaryT works for 0 and non-0.

We should thus review that other scripts still work.  The new momentaryT
has been added for those controls that actually are binary but behave
always temporarily when mapped to Mixxx.  We should review these one
by one with a controller different than Xponent, because Xponent sends
non-zero note offs which fucks up everything.

Note also that now it seems that the existence of momentaryT and
option.switch provide very similar functionality.  The rule of thumb
is to use the first when it is a property of the behaviour itself, and
the later when it is a customization.