Re-implement the fix for Xponent note offs in Xponent...
...we do so by chaning the behaviour option API such that it can proccess
the whole MIDI event.  We use this chance to simplify the transform API
too such that it is equivalent to that defined in mixco.transform for
non-linear transforms.

Note that we could also implement momentaryT in the same way that Mixxx
seem to does this -- i.e. applying "not current value" on note-off.
However, doing so scares me because it means that when releasing a
modifier while a button that depends on it is held, we could get into
an inverted state that is very annoying for the user to fix, while
otherwise she can just press the button again.