Prepare package for usage via NPM
Before, it was assumed that one was working in the `mixco` project root
folder and interoperate via the `make`.  We want to move towards a
standalone workflow, where cloning the `git` sources is not needed.

This commit is the first step.  After installing the library via NPM,
one can use `require('mixco/lib/[MODULE]')` to import mixco modules.

Also the `make` output is a bit less noisy now.