Generate test coverage for the CoffeeScript code
Before we generated coverage for the JavaScript generated, which is not
that useful.  Now we generate coverage for the original Coffee sources.

First, we had to change our testing framework to Mocha + Chai + Sinon.
This is so because coffee-istanbul does not work with jasmine-node.

Also, to avoid generating coverage for the generated code, we had to change
the imports in the example scripts to include directly the Coffee sources.
While this makes them slightly less portable and less useful as an
example, this is probably alright.