Fix generating the Mixxx scripts
This was broken in:
    c445abb Generate test coverage for the CoffeeScript code

The problem is that now we include the coffee-script original sources
directly in the scripts, such that the code coverage properly works.
Browserify was getting confused, because we did not instruct it how to
deal with coffee-script sources.

Another problem is that latest Browserify would not generate a named
module for the entry point even when the right --require parameter is
passed.  This has been solved:

  https://github.com/substack/node-browserify/issues/1021