Enforce .mixco.* extension for scripts and automatically define 'name'
Before, the script authors had to define `name` attribute in their
scripts and make sure it matches the script file minus extension.  This
is not needed anymore.

Also, as already introduced in the documentation for the CLI, scripts
should have the extension .mixco.*.  This should make clearer the
idiosincracies of the script.