Take into account note_on/off for pressed state
This is how Mixxx 2.0 does it.  Which is the right thing and avoids
further hacks.  For this, we synthesize a `pressed` attribute for the
MIDI event, that depends on the value and the midi status.  We needed to
change the transform API, instead of just taking the value, it now takes
the MIDI event, since binary transforms need to look at the pressed
state.  This is now less symmetric though.